Remove the need to call factory function
