--- enterprise/frontend/src/metabase-enterprise/custom_viz/custom-viz-props.ts
@@ -1,15 +1,0 @@
-import type { CreateCustomVisualizationProps } from "custom-viz/src/types/viz";
-
-interface BuildCustomVizPropsOptions {
-  locale: string;
-  getAssetUrl: (assetPath: string) => string;
-}
-
-export function buildCustomVizProps(
-  opts: BuildCustomVizPropsOptions,
-): CreateCustomVisualizationProps {
-  return {
-    locale: opts.locale,
-    getAssetUrl: opts.getAssetUrl,
-  };
-}
