+++ frontend/src/custom-viz/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./column-types";
 export * from "./format";
-export * from "./lib";
 export * from "./measure-text";
 export * from "./types";
