--- frontend/src/custom-viz/src/lib/index.ts
@@ -1,1 +1,0 @@
-export * from "./createDefineSetting";
